requirements.txt
requests==2.31.0
Pillow==10.0.1
Flask==3.0.0
Werkzeug==3.0.1
pygame==2.5.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
tensorflow==2.13.0
joblib==1.3.2
textual==0.44.1
rich==13.7.0
